Api`s crud implementation

+++ TodoApi/TodoApi/Controllers/TodoController.cs
@@ -6,10 +6,10 @@ using System.Security.Claims;
 using Todo.DataService;
 using Todo.DataService.DataServices;
 using Todo.Domain;
+using TodoApi.Models;
 
 namespace TodoApi.Controllers
 {
-    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class TodoController : ControllerBase
@@ -44,9 +44,7 @@ namespace TodoApi.Controllers
                 return NotFound("Todo couldn't be found.");
             }
 
-            var email = User.Identity.Name;
-
-            var user = _userDataService.GetByEmail(email);
+            var user = _userDataService.GetByEmail(User.Identity.Name);
 
             if(user.UserRole == 1)
             {
@@ -54,7 +52,7 @@ namespace TodoApi.Controllers
             }
             else
             {
-                if (_userDataService.CanUserViewTodo(user, todo))
+                if (_userDataService.CanUserManageTodo(user, todo))
                 {
                     return Ok(todo);
                 }
@@ -67,21 +65,31 @@ namespace TodoApi.Controllers
         }
         // POST: api/Todo
         [HttpPost]
-        public IActionResult Post([FromBody] Todo.Domain.Todo todo)
+        [Authorize(Policy = "ValidAccessToken")]
+        public IActionResult Post([FromBody] TodoPostModel todo)
         {
+            var user = _userDataService.GetByEmail(User.Identity.Name);
+            if (user.UserRole == 1)
+            {
+                return BadRequest("Administrator cannot create new todo");
+            }
+
             if (todo == null)
             {
                 return BadRequest("Todo is null.");
             }
-            _dataService.Add(todo);
+
+            var newTodo = new Todo.Domain.Todo { Text = todo.Text, IsDone = false, TodoListId = _userDataService.GetUsersTodoListId(_userDataService.GetByEmail(User.Identity.Name)) };
+            _dataService.Add(newTodo);
             return CreatedAtRoute(
-                  "Get",
-                  new { Id = todo.TodoId },
+                  "GetTodo",
+                  new { Id = newTodo.TodoId },
                   todo);
         }
         // PUT: api/Todo/5
         [HttpPut("{id}")]
-        public IActionResult Put(long id, [FromBody] Todo.Domain.Todo todo)
+        [Authorize(Policy = "ValidAccessToken")]
+        public IActionResult Put(long id, [FromBody] TodoUpdateModel todo)
         {
             if (todo == null)
             {
@@ -92,11 +100,24 @@ namespace TodoApi.Controllers
             {
                 return NotFound("The Todo record couldn't be found.");
             }
-            _dataService.Update(todoToUpdate, todo);
-            return NoContent();
+
+            var user = _userDataService.GetByEmail(User.Identity.Name);
+
+            if (_userDataService.CanUserManageTodo(user, todoToUpdate))
+            {
+                _dataService.Update(todoToUpdate, new Todo.Domain.Todo { Text = todo.Text, IsDone = todo.IsDone });
+                return NoContent();
+            }
+            else
+            {
+                return Unauthorized();
+            }
+
+
         }
         // DELETE: api/Todo/5
         [HttpDelete("{id}")]
+        [Authorize(Policy = "ValidAccessToken")]
         public IActionResult Delete(long id)
         {
             Todo.Domain.Todo todo = _dataService.Get(id);
@@ -104,8 +125,28 @@ namespace TodoApi.Controllers
             {
                 return NotFound("The Todo record couldn't be found.");
             }
-            _dataService.Delete(todo);
-            return NoContent();
+
+            var user = _userDataService.GetByEmail(User.Identity.Name);
+
+
+            if (user.UserRole == 1)
+            {
+                _dataService.Delete(todo);
+                return NoContent();
+            }
+            else
+            {
+                if (_userDataService.CanUserManageTodo(user, todo))
+                {
+                    _dataService.Delete(todo);
+                    return NoContent();
+                }
+                else
+                {
+                    return Unauthorized();
+                }
+            }
+
         }
     }
 }
